Use new AbstractEntityEditDetailsPane
OPEN - task 64: Create CRUD functionality for Brand entity 
http://github.com/DevOpsDistilled/OpERP/issues/issue/64
